// src/models/Orders.ts
export interface Orders {
    customer_address: string;
    customer_avatar_path: string;
    customer_birth_date: string;
    customer_gender: string;
    customer_name: string;
    customer_phone_number: string;
    orders_created_at: string;
    orders_customer_id: number;
    orders_id: number;
    orders_price: number;
    orders_product_id: number;
    orders_quantity: number;
    orders_status: string;
    product_color: string;
    product_description: string;
    product_name: string;
    product_preview_url: string;
    product_price: number;
    product_rate: number;
    product_type: string;
}